requirements.txt
pymongo
gitpython